Add a nuspec and script to generate the nuget package.
Run `nugetpack` in the BarcodeLib directory to produce a publishable
nuget package. This script wraps calling msbuild and nuget while
setting Configuration=Release. The nuspec has some magic to reduce the
chance that anyone accidentally invokes `nuget pack` directly.

The files Release Notes.txt and LICENSE are distributed in the nuget
package but are stored when the built assembly. To prevent nuget from
packaging Release Notes.txt in such a way that the file would be
copied into the source tree of any project adding the package via
nuget, its build type was changed from Content to None. Not the best
link, but http://stackoverflow.com/a/17300437/429091 .