Add dynamic barcode Width/Height sizing calculations for issue #5.
The new nullable BarWidth property allows the user to specify the
desired bar width in pixels. If set, the necessary Width will be
calculated when drawing the image and stored into the Width property.

The new nullable AspectRatio property allows the user to specify that
the Height should be dynamically calculated from the Width. If set,
Height will be set to Width divided by AspectRatio when drawing the
image and stored into the Height property.

These features enable application developers to support encoding
strings with unconstrainted lengths into barcodes. It also allows
application developers to avoid unwanted whitespace to the left and
right of a barcode if desired.