Code128: Remove Code128Entry struct.
This should be slightly more memory efficient. Uses the index of the
list as the value instead of storing the index in the element
itself. Reverse lookup of the symbolic name of a particular barcode
pattern is replaced with dictionary key probing which happens for the
initial character and when switching codes.

See #123, addresses comments by @jangernert.