BarcodeStandardExample: Use implicit package references.
By setting RestoreProjectStyle to PackageReference in
Directory.Build.props, the entire solution is opted into using
PackageReference-style nuget restore, even if those packages do not
themselves have any direct PackageReference. This means that any
PackageReference to another project will automatically pull in the
nuget packages of that other project. For BarcodeStandardExample, this
means that it will automatically get the version of
System.Drawing.Common used by BarcodeStandard.

This may be a little sloppy, in that removing a PackageReference from
BarcodeStandard may break the build of BarcodeStandardExample. But I
think the simpler project files and reduced number of
manually-maintained PackageReference is worth it. And such a break
should be caught by CI.

This gets rid of packages.config usage. The packages.config style of
referencing nugets is old-style and the very fact that this puts
Reference and HintPath in the .csproj is unfortunate. So it is good to
move away from that.